+++ packages/compiler/examples/source-ts/dist-types/utils/nested/check.d.ts
@@ -1,0 +1,1 @@
+export declare const check: (str: string) => string | false;