+++ packages/starter/templates/addons/tester/library/jest.extend.js
@@ -1,0 +1,3 @@
+module.exports = {
+  resolver: 'jest-ts-webcompat-resolver',
+};